target/i386/kvm: get/set GP registers of TD during migration
The GP registers are emulated by KVM when TD executes a hypercall (i.e.
TDG.VP.VMCALL). Live migration could pause a TD when it's in KVM side
execution of the hypercall, and in this case, the GP registers need to
be migrated to the destination before resume the guest to run.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>