i386/tdx: add a new tdvf section TDVF_SECTION_TYPE_PAYLOAD
TDVF_SECTION_TYPE_PAYLOAD is defined in the TDVF spec as a piece of
memory to be added by VMM via TDH.MEM.ADD. As documented in the TDVF spec,
it is used by TD-Shim to load the service TD core (e.g. the MigTD
excutable). Add the section type to allow QEMU to pass the section
memory to KVM via the KVM_TDX_INIT_MEM_REGION command to add the pages
to TD via TDH.MEM.ADD.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>