migration/ram: save a private page
If a page is a private page, calls the cgs framework API to save the
page via the vendor specific implementation. Otherwise, send the page
via the legacy code path, where ram_save_target_page_legacy is renamed
to ram_save_target_page_shared to indicate the path is for sending
shared pages.

Add the cgs-private-pages counter to help users understand the total
amount of private pages that have been sent, which could be userful when
doing performance tuning.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>