kvm: add kvm_gmem_default_shared
The kvm_gmem_default_shared flag control whether the gmem pages are
initialized as shared or private pages. For the TDX case, a migratable
TD guest requires the gmem pages to be shared on initialization so that
guest's convert of shared pages to private pages can be captured by
userspace. For the destination TD guest during live migration, its gmem
pages are initialized as private as KVM pre-fualts all the gmem pages
as private and shared pages will be converted from private during the
migration flow.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>