migration: Cancel live migration when a private-to-shared page conversion occurs
When a private-to-shared page conversion occurs during live migration,
cancel the migration process. This behavior is consistent with how changes
in RAM size are handled during legacy VM live migrations.

Support of such conversions during live migration is a bit complex and
tricky, so let's consider it for future optimizations later.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>